Add all-bytes and char-from-byte
A couple of helpers that we're about to use to solve the challenge. Put
in helpers because they're probably generic and useful elsewhere.

all-bytes is simply a sequence of all possible byte combinations

char-from-byte converts, as you would expect, a byte to a character,
taking into account that bytes are signed and their values may need to
be coerced into the valid range for char.